add api

See merge request s07-webmobile1-sub2/S07P12A304!14

+++ BackEnd/src/main/java/com/ssafy/db/entity/UserClass.java
@@ -22,6 +22,6 @@ public class UserClass {
     private Classes classes;
 
     @Column
-    int studentId;
+    Long studentId;
 
 }
